Add files via upload
The Hangman game is a word-guessing game. The program randomly selects a word from a predefined list, and the player's goal is to guess the letters of the word one by one. The player has a limited number of attempts to guess the correct letters, after each correct guess  the game continues until the player correctly guesses the word or runs out of attempts.